read_flash interface consistency
Everywhere else we accept an uint32_t for numWords, so do it here, too.

This commit also changes flash writes to be chunked in groups of 256
bytes (64 32-bit words), consistent with the spec. We're not actually
sure if this is necessary.

--BRJ

+++ hardware/APS/libaps2-cpp/src/APS/APS2.h
@@ -96,7 +96,7 @@ public:
 
 	//Flash read/write
 	int write_flash(const uint32_t &, vector<uint32_t> &);
-	vector<uint32_t> read_flash(const uint32_t &, const uint16_t &);
+	vector<uint32_t> read_flash(const uint32_t &, const uint32_t &);
 
 	//CLPD DRAM
 	int write_bitfile(const uint32_t &, const string &);
